cambios en modelos

+++ src/main/java/com/spootify/backend_spootify/Models/Merch.java
@@ -1,7 +1,6 @@
 package com.spootify.backend_spootify.Models;
 
 import java.sql.Clob;
-import java.util.List;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -10,7 +9,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -37,7 +35,7 @@ public class Merch {
 
     private float precio_merch;
 
-    private float stock_merch;
+    private int stock_merch;
 
     private int id_artista;
